refactor: extract connection and history helpers
